pyproject.toml
[tool.poetry]
name = "discordbot"
version = "0.1.0"
description = ""
authors = ["Pavel Zubarev <pavelzub1996@gmail.com>", "Alexey Kunitsky <kunitsky.alexey@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.20.0"
python-opendota = "^1.0.2"
firebase-admin = "^5.2.0"
py-cord = "2.0.0rc1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
